backend/requirements_supabase.txt
# Requirements para Supabase - DEPENDENCIAS MÍNIMAS Y FUNCIONALES
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pydantic==2.5.0
